felex.py: C/C++ comment lines in .feature files parsed...
The comment lines are recognized only when placed on separate
lines. That is, C-comment must start on the line or end on the line.
The lines in between are just 'line' -- must be solved by syntactic
parser.

Whitespaces before the comment or after the comment are ignored.
Whitespace inside the comment are part of the comment text.